[sync] split tests
