+++ src/sync/test.js
@@ -3,35 +3,7 @@ import clone from 'lodash.clonedeep'
 import { mockDatabase } from '../__tests__/testModels'
 
 import { fetchLocalChanges, markLocalChangesAsSynced, applyRemoteChanges } from './impl'
-import { addToRawSet, setRawColumnChange } from './helpers'
 import { resolveConflict, prepareCreateFromRaw } from './syncHelpers'
-
-describe('addToRawSet', () => {
-  it('transforms raw set', () => {
-    expect(addToRawSet('', 'foo')).toBe('foo')
-    expect(addToRawSet('foo', 'bar')).toBe('foo,bar')
-    expect(addToRawSet('foo,bar', 'baz')).toBe('foo,bar,baz')
-    expect(addToRawSet('foo,bar', 'foo')).toBe('foo,bar')
-    expect(addToRawSet('foo,bar', 'bar')).toBe('foo,bar')
-  })
-})
-
-describe('setRawColumnChange', () => {
-  it('adds to _changed if needed', () => {
-    const test = (input, column, output) => {
-      const raw = { ...input }
-      setRawColumnChange(raw, column)
-      expect(raw).toEqual(output)
-    }
-    test({ _status: 'synced', _changed: '' }, 'foo', { _status: 'updated', _changed: 'foo' })
-    test({ _status: 'created', _changed: '' }, 'foo', { _status: 'created', _changed: '' })
-    test({ _status: 'updated', _changed: '' }, 'foo', { _status: 'updated', _changed: 'foo' })
-    test({ _status: 'updated', _changed: 'foo,bar' }, 'bar', {
-      _status: 'updated',
-      _changed: 'foo,bar',
-    })
-  })
-})
 
 describe('Conflict resolution', () => {
   it('can resolve per-column conflicts', () => {
